requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
astroid==4.0.4
attrs==26.1.0
audioop-lts==0.2.2
certifi==2026.2.25
charset-normalizer==3.4.7
dill==0.4.1
discord.py==2.7.1
frozenlist==1.8.0
idna==3.11
isort==8.0.1
mccabe==0.7.0
multidict==6.7.1
platformdirs==4.9.6
propcache==0.4.1
pylint==4.0.5
python-dotenv==1.2.2
requests==2.33.1
tomlkit==0.14.0
urllib3==2.6.3
yarl==1.23.0
